WiP: endpoint project
