+++ somsolet_api/serializer/__init__.py
@@ -1,1 +1,2 @@
 from .campaign import CampaignSerializer
+from .project import ProjectSerializer
