+++ somsolet_api/urls.py
@@ -1,9 +1,11 @@
 from django.urls import include, path
 from rest_framework import routers
-from .views import CampaignViewSet
+from .views import (CampaignViewSet, ProjectViewSet)
 
 router = routers.DefaultRouter()
 router.register(r'campaign', CampaignViewSet)
+router.register(r'project', ProjectViewSet)
+
 
 # Wire up our API using automatic URL routing.
 # Additionally, we include login URLs for the browsable API.
