+++ somsolet_api/views/__init__.py
@@ -1,1 +1,2 @@
 from .campaign import CampaignViewSet
+from .project import ProjectViewSet
